+++ front/app/components/Button/Button.module.css
@@ -1,0 +1,16 @@
+.button {
+  background-color: #971805;
+  color: #fbf6ee;
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 0.25rem;
+  cursor: pointer;
+  font-size: 24px;
+  font-weight: bold;
+}
+
+.button:disabled {
+  color: #333;
+  background-color: #ffffff;
+  border: 2px solid #333;
+}
